feat: Add service worker registration
This commit adds code to register a service worker in the main.tsx file. The service worker is registered based on the environment mode, either using "/sw/worker.js" for production or "/dev-sw.js?dev-sw" for development. The type of the registration is also determined by the environment mode, with "classic" used for production and "module" used for development.