Add makeBoard function and import it in worker.ts
This commit adds a new file `makeBoard.ts` which contains the `makeBoard` function. The function creates a new instance of the `Board` class from the "@google-labs/breadboard" package and sets up input and output connections. The `makeBoard` function is then imported in `worker.ts`, replacing the previous implementation of the same function.