Add TOP_STORIES flag for retrieving top stories
This commit adds a new flag, TOP_STORIES, which can be used to control whether the code retrieves top stories from Hacker News. If the flag is set to true, the code will fetch the top story IDs and wire them to the popStory list. Otherwise, this functionality is skipped. This change improves flexibility and allows for easier testing of different scenarios.